Clean code 🪒
A few file organization for better code redability
