+++ go.mod
@@ -1,0 +1,3 @@
+module simplehttp
+
+go 1.22.0
